Rework activity planner to allow splitting, filling remaining time
With the old logic, when activities wouldn't fit in a timeslot they
would be pushed into the next day.

The new logic can split a particularly large activity so that timeslots
get filled. Once both slots are filled the remaining activities will be
moved in front of the evening slot, starting increasingly earlier if
necessary.