Rework activity planner to have variable day model, fix warning on toTime
Originally it would assume fixed hours, now some of the quiet hour
intervals (end of morning quiet hours, start of evening quiet hours) can
be changed by putting a _planning.tsv file in the input folder.

This commit also fixes an issue in the controller where passing a toTime
which falls on the next monday of the week would raise an HTTP 400
response and a warning in the logs. The code treated the toTime as
inclusive, whereas ManicTime simply assumes that it's the exclusive end
offset. The code will now subtract 24 hours from this time offset to
calculate the last included day.